requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
discord==2.1.0
discord.py==2.1.0
frozenlist==1.3.3
future==0.18.2
idna==3.4
multidict==6.0.4
Pillow==9.4.0
redis==4.4.0
requests==2.28.1
six==1.16.0
spotipy==2.22.0
support-developer==1.0.5
toml==0.10.2
update-checker==0.18.0
urllib3==1.26.13
websocket-client==1.4.2
yarl==1.8.2
